+++ controller/editorapicontroller.php
@@ -316,6 +316,12 @@ class EditorApiController extends OCSController {
 		$restrictedEditing = false;
 		$fileStorage = $file->getStorage();
 		if (empty($shareToken) && $fileStorage->instanceOfStorage("\OCA\Files_Sharing\SharedStorage")) {
+
+			$params["editorConfig"]["coEditing"] = [
+				"mode" => "strict",
+				"change" => false
+			];
+
 			$storageShare = $fileStorage->getShare();
 			if (method_exists($storageShare, "getAttributes")) {
 				$attributes = $storageShare->getAttributes();