+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ## Added
 - reference data from coediting
 - selecting a document to combine from the storage
+- opening a reference data source
 
 ## 8.2.3
 ## Added