+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - selecting a document to combine from the storage
 - Arabic and Serbian empty file templates
 - opening a reference data source
+- changing a reference data source
 
 ## 8.2.3
 ## Added