+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changed
 - compatible with ownCloud Web 7.0
 - list of users to protect ranges of cells
+- offline viewer for share link
 
 ## Added
 - reference data from coediting