+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - compatible with ownCloud Web 7.0
 - list of users to protect ranges of cells
 - offline viewer for share link
+- updatable list of supported formats
 
 ## Added
 - reference data from coediting